requirements/base.txt
Django==2.1.1
django-configurations==2.1
django-environ==0.4.5
djangorestframework==3.8.2
